fix: naming and structure refactoring
